chore(build): add git commit hash and expect nightly releases

+++ appimage/build-appimage.sh
@@ -29,6 +29,13 @@
 
 readonly DEBUG="$1"
 
+# Set this to True to upload the PR version of the 
+# AppImage to transfer.sh for testing.
+readonly UPLOAD_PR_APPIMAGE="True"
+
+# Git commit to use for nightlies and PR upload.
+readonly GIT_HASH=$(git rev-parse --short HEAD)
+
 # Fail out on error
 set -exo pipefail
 
@@ -44,6 +51,7 @@ if [ ! -f ./appimage/build-appimage.sh ]; then
 fi
 
 mkdir -p ./output
+mkdir -p ./nightly-appimage # AppImages specially made for nightly release.
 
 if [ "$DEBUG" == "Debug" ]
 then
@@ -57,9 +65,12 @@ then
         /bin/bash
 else
     docker run --rm \
+        -e GIT_HASH \
+	-e CIRP_GITHUB_REPO_SLUG \
         -e TRAVIS_TAG \
         -v $PWD:/qtox \
         -v $PWD/output:/output \
+        -v $PWD/nightly-appimage:/nightly-appimage \
         debian:stretch-slim \
         /bin/bash -c "/qtox/appimage/build.sh"
 fi
@@ -80,10 +91,14 @@ then
 
     sha256sum "$OUTFILE" > "$OUTFILE".sha256
 else
-    # upload PR builds to test them.
-    echo "Uploading AppImage to transfer.sh"
-    curl --upload-file "./output/qTox-x86_64.AppImage" "https://transfer.sh/qTox-x86_64.AppImage" > ./upload
-    echo "$(cat ./upload)"
-    echo -n "$(cat ./upload)\\n" >> ./uploaded-to
-    rm -rf ./upload ./uploaded-to
+    if [ "$UPLOAD_PR_APPIMAGE" == "True" ] 
+    then
+        # upload PR builds to test them.
+        echo "Uploading AppImage to transfer.sh"
+        curl --upload-file "./output/qTox-$GIT_HASH-x86_64.AppImage" "https://transfer.sh/qTox-$GIT_HASH-x86_64.AppImage" > ./upload
+        echo "$(cat ./upload)"
+        echo -n "$(cat ./upload)\\n" >> ./uploaded-to
+        rm -rf ./upload ./uploaded-to
+	sha256sum "./output/qTox-$GIT_HASH-x86_64.AppImage"
+    fi
 fi
